backend/requirements.txt
autopep8==1.6.0
Django==2.2.19
django-cors-headers==3.10.0
django-extra-fields==3.0.2
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
flake8==4.0.1
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.8.6
pytils==0.3
